safari-534.53-branch/LayoutTests/platform/mac/accessibility/selection-value-changes-for-aria-textbox-expected.txt

ALERT: Successfully received AXSelectedTextChanged
ALERT: Successfully received AXSelectedTextChanged
ALERT: Successfully received AXSelectedTextChanged
ALERT: Successfully received AXSelectedTextChanged
ALERT: Successfully received AXSelectedTextChanged
ALERT: Successfully received AXValueChanged
conhello tent inside
This tests that ARIA removing an element from a live region sends the correct notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS successfullyParsed is true

TEST COMPLETE
